# apps/api/src/main/java/fr/glowjob/hackathon/web/france_travail/model/offers/reference/ContractTypeReference.java
package fr.glowjob.hackathon.web.france_travail.model.offers.reference;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
public class ContractTypeReference extends AbstractReference {
}